Implementation of Dijkstra's Algorithm
Dijkstra's algorithm maintains the minimum edge that connects two sets of
vertices. This minimum edge is compared with the existing minimum path to it's
head vertex and updates the minimum path distance if it is a new minimum. This
algorithm runs in O(m log n) but must have no negative edge weights in the
graph.